HUE-1777 [beeswax] Unify is_auto status of the queries
Example queries not modified are now seen as non auto.
Showing all queries by defaut.
Auto queries are not shown as Saved queries anymore in My queries.
The UX is bad, we should have 2 sections with radio on/off for auto query
and see all queries.
Tested with search and flipping auto/non auto.

Do not remove the header row in a gzip file in create table.