HUE-1658 [oozie] Cron like scheduling
Replace old double dropdown by an text crontab.
Try to go convert old frequency format to the new crontab one.
Adding a new coordinator util lib as it needs to work on both edit and create
coordinator tab.

Dashboard should detect if the coordinator is using the crontab format and
try display it in english with the plugin (todo in another jira)
Choice was made to remove the previous form. Frequencies like every 3 days
is a bit more complicated are not supported by the plugin and some like
every 2 days are not supported by cron AFAIK.

Maybe add the advanced options to use the old dropdown.